+++ libs/adapter/nest/message/message.controller.spec.ts
@@ -3,7 +3,7 @@ import { MessageController } from './message.controller';
 
 describe('MessageController', () => {
   const handler = { handle: jest.fn() };
-  const logger = { log: jest.fn() };
+  const loggingService = { log: jest.fn() };
 
   let controller: MessageController;
 
@@ -11,8 +11,8 @@ describe('MessageController', () => {
     const module = await Test.createTestingModule({
       controllers: [MessageController],
       providers: [
+        { provide: 'Context', useValue: { loggingService } },
         { provide: 'HandlerMap', useValue: { Test: handler } },
-        { provide: 'Logger', useValue: logger },
       ],
     }).compile();
 
@@ -26,8 +26,8 @@ describe('MessageController', () => {
       const message = { type: 'Test', by: 'by', timestamp: 'timestamp' };
       await controller.postMessage(message);
 
-      expect(logger.log).toHaveBeenCalledTimes(1);
-      expect(logger.log).toHaveBeenNthCalledWith(1, message);
+      expect(loggingService.log).toHaveBeenCalledTimes(1);
+      expect(loggingService.log).toHaveBeenNthCalledWith(1, 'message', message);
 
       expect(handler.handle).toHaveBeenCalledTimes(1);
       expect(handler.handle).toHaveBeenNthCalledWith(1, message);